Refactoring to get closer to custom transforms
This change moves the radius orbit transform to its own module, as well
as the GIF encoding things and the party palette. The aim is to golf
this down a bit to some isolated components so that we can extend things
with additional transforms, as well as consider adding tests.